IPC-Rückgabewerte
* Rückgaben von Popupthreads sind jetzt uintmax_t anstatt von uintptr_t.

+++ src/lib/common/system/popup_handler.c
@@ -1,17 +1,17 @@
 #include <ipc.h>
 
 
-extern uintptr_t (*popup_entries[])(uintptr_t);
+extern uintmax_t (*popup_entries[])(uintptr_t);
 
 
-void popup_message_handler(int index, uintptr_t (*handler)(void))
+void popup_message_handler(int index, uintmax_t (*handler)(void))
 {
     // FIXME: Überlauf und so
-    popup_entries[index] = (uintptr_t (*)(uintptr_t))handler;
+    popup_entries[index] = (uintmax_t (*)(uintptr_t))handler;
 }
 
 
-void popup_shm_handler(int index, uintptr_t (*handler)(uintptr_t))
+void popup_shm_handler(int index, uintmax_t (*handler)(uintptr_t))
 {
     popup_entries[index] = handler;
 }
